Introduce `rustpython-format` crate
