+++ format/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "rustpython-format"
+edition = "2021"
+version = "0.2.0"
+description = "Format helpers for RustPython"
+authors = ["RustPython Team"]
+repository = "https://github.com/RustPython/Parser/"
+license = "MIT"
+
+[dependencies]
+rustpython-literal = { workspace = true }
+
+bitflags = "2.2.1"
+itertools = "0.10.5"
+num-bigint = { workspace = true }
+num-traits = { workspace = true }
