+++ format/src/cformat.rs
@@ -1,9 +1,9 @@
 //! Implementation of Printf-Style string formatting
 //! as per the [Python Docs](https://docs.python.org/3/library/stdtypes.html#printf-style-string-formatting).
-use crate::{float, format::Case};
 use bitflags::bitflags;
 use num_bigint::{BigInt, Sign};
 use num_traits::Signed;
+use rustpython_literal::{float, format::Case};
 use std::{
     cmp, fmt,
     iter::{Enumerate, Peekable},
