+++ literal/Cargo.toml
@@ -4,16 +4,13 @@ version = "0.2.0"
 description = "Common literal handling utilities mostly useful for unparse and repr."
 authors = ["RustPython Team"]
 edition = "2021"
-repository = "https://github.com/RustPython/RustPython"
+repository = "https://github.com/RustPython/Parser/"
 license = "MIT"
 
 [dependencies]
-bitflags = "2.2.1"
-itertools = "0.10.5"
-num-bigint = { workspace = true }
-num-traits = { workspace = true }
 hexf-parse = "0.2.1"
 lexical-parse-float = { version = "0.8.0", features = ["format"] }
+num-traits = { workspace = true }
 unic-ucd-category = "0.9"
 
 [dev-dependencies]
