+++ literal/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod cformat;
 pub mod char;
 pub mod escape;
 pub mod float;
